+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add Topology Discovery
 - Add Support for StartupProbe
 - Add ARM64 support for Operator Docker image
+- Add ALPHA Rebalancer support
 
 ## [1.2.4](https://github.com/arangodb/kube-arangodb/tree/1.2.4) (2021-10-22)
 - Replace `beta.kubernetes.io/arch` Pod label with `kubernetes.io/arch` using Silent Rotation